refactor: exams와 questions 도메인 분리
