+++ src/main/java/capstone/examlab/exams/controller/ExamsController.java
@@ -1,8 +1,6 @@
 package capstone.examlab.exams.controller;
 
 import capstone.examlab.exams.dto.ExamList;
-import capstone.examlab.exams.dto.QuestionsList;
-import capstone.examlab.exams.dto.QuestionsOption;
 import capstone.examlab.exams.service.ExamsService;
 import capstone.examlab.exams.dto.ExamType;
 import lombok.RequiredArgsConstructor;
@@ -13,7 +11,7 @@ import org.springframework.web.bind.annotation.*;
 @Slf4j
 @RequiredArgsConstructor
 @RestController
-@RequestMapping("/exams")
+@RequestMapping("api/v1/exams")
 public class ExamsController {
 
     private final ExamsService examsService;
@@ -28,11 +26,5 @@ public class ExamsController {
     public ExamType getExamType(@PathVariable("id") Long id) {
         ExamType examType = examsService.getExamType(id);
         return examType;
-    }
-
-    @GetMapping("/{id}/questions")
-    public QuestionsList getExamQuestions(@PathVariable("id") Long id, @ModelAttribute QuestionsOption questionsOption) {
-        log.info("questionOptionDto = {}", questionsOption);
-        return examsService.getQuestionsList(id, questionsOption);
     }
 }
