+++ src/main/java/capstone/examlab/exams/service/ExamServiceImpl.java
@@ -55,82 +55,4 @@ public class ExamServiceImpl implements ExamsService {
 
         return examList;
     }
-
-    public QuestionsList getQuestionsList(Long id, QuestionsOption questionsOption) {
-        //새로운 시험지 생기면 조건 추가
-        if(id == 1) return questionsService.findByDriverLicenseQuestions(questionsOption);
-        return null;
-    }
-
-   /* @Override
-    public QuestionsList getQuestionsList(QuestionsOption questionsOption) {
-        QuestionsList questionsList = new QuestionsList();
-
-        questionsList.add(Question.builder()
-                .id(1L)
-                .type("객관식")
-                .question("1번 문제")
-                .questionImageUrls(new ArrayList<String>() {{
-                    add("1번 문제 이미지");
-                }})
-                .questionImageDescriptions(new ArrayList<String>() {{
-                    add("1번 문제 이미지 설명");
-                }})
-                .options(new ArrayList<String>() {{
-                    add("1번 답");
-                    add("2번 답");
-                    add("3번 답");
-                    add("4번 답");
-                }})
-                .answers(new ArrayList<Integer>() {{
-                    add(1);
-                    add(3);
-                }})
-                .explanation("1번 문제 설명")
-                .explanationImageUrls(new ArrayList<String>() {{
-                    add("1번 문제 설명 이미지");
-                }})
-                .explanationImageDescriptions(new ArrayList<String>() {{
-                    add("1번 문제 설명 이미지 설명");
-                }})
-                .tags(new ArrayList<String>() {{
-                    add("표지");
-                }})
-                .build());
-
-        questionsList.add(Question.builder()
-                .id(2L)
-                .type("객관식")
-                .question("2번 문제")
-                .questionImageUrls(new ArrayList<String>() {{
-                    add("2번 문제 이미지");
-                }})
-                .questionImageDescriptions(new ArrayList<String>() {{
-                    add("2번 문제 이미지 설명");
-                }})
-                .options(new ArrayList<String>() {{
-                    add("1번 답");
-                    add("2번 답");
-                    add("3번 답");
-                    add("4번 답");
-                }})
-                .answers(new ArrayList<Integer>() {{
-                    add(2);
-                    add(4);
-                }})
-                .explanation("2번 문제 설명")
-                .explanationImageUrls(new ArrayList<String>() {{
-                    add("2번 문제 설명 이미지");
-                }})
-                .explanationImageDescriptions(new ArrayList<String>() {{
-                    add("2번 문제 설명 이미지 설명");
-                }})
-                .tags(new ArrayList<String>() {{
-                    add("상황");
-                }})
-                .build());
-
-
-        return questionsList;
-    }*/
 }
