+++ src/main/java/capstone/examlab/exams/service/ExamsService.java
@@ -2,14 +2,10 @@ package capstone.examlab.exams.service;
 
 import capstone.examlab.exams.dto.ExamList;
 import capstone.examlab.exams.dto.ExamType;
-import capstone.examlab.exams.dto.QuestionsList;
-import capstone.examlab.exams.dto.QuestionsOption;
 
 public interface ExamsService {
 
     public ExamType getExamType(Long id);
 
     public ExamList getExamList();
-
-    public QuestionsList getQuestionsList(Long id, QuestionsOption questionsOption);
 }
