+++ src/main/java/capstone/examlab/questions/dto/Image.java
@@ -1,0 +1,10 @@
+package capstone.examlab.questions.dto;
+
+import lombok.Data;
+
+@Data
+public class Image {
+    private String url;
+    private String description;
+    private String attribute;
+}
