+++ src/main/java/capstone/examlab/questions/dto/QuestionsList.java
@@ -1,4 +1,4 @@
-package capstone.examlab.exams.dto;
+package capstone.examlab.questions.dto;
 
 import java.util.ArrayList;
