+++ src/main/java/capstone/examlab/questions/dto/QuestionsOption.java
@@ -1,9 +1,8 @@
-package capstone.examlab.exams.dto;
+package capstone.examlab.questions.dto;
 
 import lombok.Builder;
 import lombok.Data;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Data
