+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
 #@RequestMapping? ?? ???
-spring.mvc.servlet.path=/api/v1
+#spring.mvc.servlet.path=/api/v1
 spring.jackson.property-naming-strategy=SNAKE_CASE
 
 #DB Secruing Resoures
